feat: Add interest rate and product management for vendor-gate and vendor-huobi
- Implemented interest rate series provider in vendor-gate.
- Created product management for futures contracts in vendor-gate.
- Established terminal configuration for vendor-gate.
- Refactored vendor-huobi to utilize SQL for product and interest rate management.
- Added API client initialization in vendor-huobi.
- Introduced terminal setup for vendor-huobi.
- Enhanced product management for perpetual and spot products in vendor-huobi.
- Updated package dependencies in vendor-huobi and vendor-hyperliquid.
- Added SQL writing capabilities for vendor-hyperliquid.
- Introduced new utility function in data-ohlc for duration conversion.
- Refactored multiple vendors to improve SQL integration and data handling.
