+++ apps/vendor-binance/src/index.ts
@@ -1,2 +1,3 @@
-import './funding_rate';
+import './interest_rate';
 import './legacy_index';
+import './product';
