+++ apps/vendor-ccxt/package.json
@@ -11,6 +11,9 @@
     "@yuants/protocol": "workspace:*",
     "@yuants/data-model": "workspace:*",
     "@yuants/data-series": "workspace:*",
+    "@yuants/sql": "workspace:*",
+    "@yuants/data-product": "workspace:*",
+    "@yuants/data-ohlc": "workspace:*",
     "rxjs": "~7.5.6",
     "ccxt": "~4.2.48"
   },
