+++ apps/vendor-coinex/package.json
@@ -12,7 +12,9 @@
     "@yuants/protocol": "workspace:*",
     "@yuants/data-model": "workspace:*",
     "@yuants/utils": "workspace:*",
+    "@yuants/sql": "workspace:*",
     "@yuants/data-series": "workspace:*",
+    "@yuants/data-product": "workspace:*",
     "@yuants/data-interest-rate": "workspace:*",
     "rxjs": "~7.5.6",
     "crypto-js": "^4.2.0"
