+++ apps/vendor-ctp/package.json
@@ -12,6 +12,8 @@
     "@yuants/data-model": "workspace:*",
     "@yuants/protocol": "workspace:*",
     "@yuants/utils": "workspace:*",
+    "@yuants/sql": "workspace:*",
+    "@yuants/data-product": "workspace:*",
     "binary-parser-encoder": "~1.5.3",
     "date-fns": "~2.29.3",
     "rxjs": "~7.5.6",
