+++ apps/vendor-gate/package.json
@@ -10,6 +10,8 @@
   "dependencies": {
     "@yuants/protocol": "workspace:*",
     "@yuants/data-model": "workspace:*",
+    "@yuants/sql": "workspace:*",
+    "@yuants/data-product": "workspace:*",
     "@yuants/utils": "workspace:*",
     "@yuants/data-series": "workspace:*",
     "@yuants/data-interest-rate": "workspace:*",
