+++ apps/vendor-huobi/package.json
@@ -11,6 +11,8 @@
     "@yuants/protocol": "workspace:*",
     "@yuants/data-model": "workspace:*",
     "@yuants/utils": "workspace:*",
+    "@yuants/sql": "workspace:*",
+    "@yuants/data-product": "workspace:*",
     "@yuants/data-series": "workspace:*",
     "@yuants/data-interest-rate": "workspace:*",
     "rxjs": "~7.5.6",
