+++ apps/vendor-hyperliquid/src/terminal.ts
@@ -1,0 +1,7 @@
+import { UUID } from '@yuants/data-model';
+import { Terminal } from '@yuants/protocol';
+
+export const terminal = new Terminal(process.env.HOST_URL!, {
+  terminal_id: process.env.TERMINAL_ID || `HyperLiquid/${UUID()}`,
+  name: 'HyperLiquid',
+});
