+++ common/changes/@yuants/data-ohlc/2025-06-12-21-40.json
@@ -1,0 +1,10 @@
+{
+  "changes": [
+    {
+      "packageName": "@yuants/data-ohlc",
+      "comment": "add util",
+      "type": "minor"
+    }
+  ],
+  "packageName": "@yuants/data-ohlc"
+}
