+++ common/changes/@yuants/vendor-bitget/2025-06-12-21-40.json
@@ -1,0 +1,10 @@
+{
+  "changes": [
+    {
+      "packageName": "@yuants/vendor-bitget",
+      "comment": "refactor to SQL",
+      "type": "patch"
+    }
+  ],
+  "packageName": "@yuants/vendor-bitget"
+}
