+++ libraries/data-ohlc/etc/data-ohlc.api.md
@@ -5,6 +5,9 @@
 ```ts
 
 // @public
+export const convertDurationToMilliseconds: (duration: string) => number;
+
+// @public
 export interface IOHLC {
     close: string;
     closed_at: string;
